components/esp-dl/test_apps/requirements.txt
idf-build-apps
pytest
pytest-embedded-idf
pytest-embedded-serial-esp
pytest-rerunfailures
pytest-timeout
toml
